# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Panditji Micro Service API Development",
        "description": "Panditji Micro Service API",
        "termsOfService": "https://panditjisamagri.com/",
        "contact": {
            "email": "satish.soni@globalspace.in"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://panditjisamagri.com/"
        }
    ],
    "paths": {
...
        "/api/v1/pandit/yajman/create": {
            "post": {
                "tags": [
                    "Yajman"
                ],
                "summary": "Register a yajman",
                "operationId": "RegisterYajman",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "yajman_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "yajman_name": {
                                        "type": "string",
                                        "example": "mahesh"
                                    },
                                    "mobile_number": {
                                        "type": "string",
                                        "example": "7499670180"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "21"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "384"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Gandhi square anjangaon bari"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "example": "2010/05/30"
                                    },
                                    "created_by": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Yajman Registration Successful"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "403"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Yajman Registration Failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pandit/getYajmans": {
            "get": {
                "tags": [
                    "Yajman"
                ],
                "summary": "Get yajman registration details",
                "operationId": "getYajmanDetails",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "Yajman_details": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pandit/view/{id}": {
            "get": {
                "tags": [
                    "Yajman"
                ],
                "summary": "Get yajman detail ",
                "operationId": "getYajmanDetailsById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "Yajman_detail": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pandit/deleteYajman/{id}": {
            "delete": {
                "tags": [
                    "Yajman"
                ],
                "summary": "delete yajman detail ",
                "operationId": "deleteYajmanById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "Yajman deleted successfully": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}